requirements.txt
yfinance
matplotlib
mplfinance
pandas
python-telegram-bot
python-dotenv
